requirements.txt
Flask
marshmallow
prometheus-client
psycopg2-binary
